数据处理/DESI_download/requirements.txt
asdf==4.2.0
asdf-astropy==0.6.1
asdf_coordinates_schemas==0.4.0
asdf_standard==1.2.0
asdf_transform_schemas==0.6.0
asdf_wcs_schemas==0.4.0
astro-datalab==2.24.0
astropy==5.3.4
astropy_healpix==1.0.3
astropy-iers-data==0.2025.5.26.0.37.21
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
chardet==5.0.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
feedparser==6.0.11
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.1
gwcs==0.21.0
httplib2==0.21.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.6.3
mimeparse==0.1.3
numpy==1.26.3
packaging==25.0
pandas==2.1.4
pillow==10.3.0
pip==25.0.1
pycurl==7.45.6
pycurl-requests==0.1.2
pyerfa==2.0.1.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyvo==1.1
PyYAML==6.0.2
requests==2.31.0
scikit-learn==1.0.2
scipy==1.11.1
semantic-version==2.10.0
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
specutils>=1.13.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.22.0
